ALSA: scarlett2: Remap Level Meter values
The values previously returned by the Level Meter control were passed
through from the interface without interpretation, but it has been
discovered that the order of the values matches the mux assignment
order which is not presented to userspace. In addition, the values for
disabled mux outputs, and mux outputs which share a source are
invalid.

This patch adds a meter_level_map array which is updated on routing
changes. The array gets used by scarlett2_meter_ctl_get() to both
present the values in a standard order and to fix up the invalid
values by zeroing them (for disabled outputs) and copy them (for mux
outputs which share a source).

Signed-off-by: Geoffrey D. Bennett <g@b4.vu>