ALSA: scarlett2: Add Focusrite Clarett 8Pre USB support
From the driver's point of view, presumably the Focusrite Clarett 8Pre
USB is identical to the Clarett+ 8Pre except for the USB ID. Add
support for this in anticipation of someone being able to test and
confirm.

If this hunch is confirmed, the USB IDs in struct
scarlett2_device_info should be removed and put in to the
scarlett2_devices[] array instead, so that the struct contents do not
need to be duplicated when two devices have identical device
information.